# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_RateItemOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RateItemOut"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RateItemOut]",
        "type": "object"
      },
      "Envelope_RateSearchOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RateSearchOut"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RateSearchOut]",
        "type": "object"
      },
      "Envelope_RateSourceOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RateSourceOut"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RateSourceOut]",
        "type": "object"
      },
...
      "Envelope_dict_str__bool__": {
        "properties": {
          "data": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[dict[str, bool]]",
        "type": "object"
      },
...
      "Envelope_list_RateSourceOut__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RateSourceOut"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[list[RateSourceOut]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "title": "Meta",
        "type": "object"
      },
...
      "RateItemOut": {
        "properties": {
          "can_edit": {
            "title": "Can Edit",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_demo": {
            "title": "Is Demo",
            "type": "boolean"
          },
          "is_expired": {
            "title": "Is Expired",
            "type": "boolean"
          },
          "item_code": {
            "title": "Item Code",
            "type": "string"
          },
          "rate": {
            "title": "Rate",
            "type": "string"
          },
          "rate_display": {
            "title": "Rate Display",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "specification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specification"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "unit_display": {
            "title": "Unit Display",
            "type": "string"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "source_label",
          "verification_status",
          "is_demo",
          "is_expired",
          "item_code",
          "description",
          "specification",
          "unit",
          "unit_display",
          "rate",
          "rate_display",
          "can_edit"
        ],
        "title": "RateItemOut",
        "type": "object"
      },
      "RateSearchOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RateItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RateSearchOut",
        "type": "object"
      },
      "RateSourceOut": {
        "properties": {
          "can_edit": {
            "title": "Can Edit",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "department": {
            "title": "Department",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_demo": {
            "title": "Is Demo",
            "type": "boolean"
          },
          "is_expired": {
            "title": "Is Expired",
            "type": "boolean"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "owned": {
            "title": "Owned",
            "type": "boolean"
          },
          "sor_name": {
            "title": "Sor Name",
            "type": "string"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "department",
          "sor_name",
          "year",
          "effective_from",
          "effective_to",
          "verification_status",
          "is_demo",
          "is_expired",
          "source_reference",
          "notes",
          "item_count",
          "owned",
          "can_edit",
          "created_at"
        ],
        "title": "RateSourceOut",
        "type": "object"
      },
...
      "SourceCreate": {
        "properties": {
          "department": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Department",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sor_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Sor Name",
            "type": "string"
          },
          "source_reference": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "state": {
            "maxLength": 80,
            "minLength": 1,
            "title": "State",
            "type": "string"
          },
          "year": {
            "maxLength": 7,
            "minLength": 7,
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "state",
          "department",
          "sor_name",
          "year",
          "effective_from"
        ],
        "title": "SourceCreate",
        "type": "object"
      },
      "SourcePatch": {
        "properties": {
          "department": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sor_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sor Name"
          },
          "source_reference": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "state": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "year": {
            "anyOf": [
              {
                "maxLength": 7,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "title": "SourcePatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__api__v1__rates__ItemCreate": {
        "properties": {
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "item_code": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Item Code",
            "type": "string"
          },
          "rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Rate"
          },
          "specification": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specification"
          },
          "unit": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "item_code",
          "description",
          "unit",
          "rate"
        ],
        "title": "ItemCreate",
        "type": "object"
      },
      "app__api__v1__rates__ItemPatch": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "item_code": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Code"
          },
          "rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          },
          "specification": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specification"
          },
          "unit": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "title": "ItemPatch",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "EstimateAI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rate-items": {
      "get": {
        "operationId": "search_items_api_v1_rate_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit"
            }
          },
          {
            "in": "query",
            "name": "include_expired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Expired",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RateSearchOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Items",
        "tags": [
          "rates"
        ]
      }
    },
    "/api/v1/rate-items/{item_id}": {
      "delete": {
        "operationId": "delete_item_api_v1_rate_items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__bool__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Item",
        "tags": [
          "rates"
        ]
      },
      "patch": {
        "operationId": "patch_item_api_v1_rate_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__rates__ItemPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RateItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Item",
        "tags": [
          "rates"
        ]
      }
    },
    "/api/v1/rate-sources": {
      "get": {
        "operationId": "list_sources_api_v1_rate_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_RateSourceOut__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sources",
        "tags": [
          "rates"
        ]
      },
      "post": {
        "operationId": "create_source_api_v1_rate_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RateSourceOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source",
        "tags": [
          "rates"
        ]
      }
    },
    "/api/v1/rate-sources/{source_id}": {
      "delete": {
        "operationId": "delete_source_api_v1_rate_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__bool__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Source",
        "tags": [
          "rates"
        ]
      },
      "patch": {
        "operationId": "patch_source_api_v1_rate_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RateSourceOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Source",
        "tags": [
          "rates"
        ]
      }
    },
    "/api/v1/rate-sources/{source_id}/items": {
      "post": {
        "operationId": "create_item_api_v1_rate_sources__source_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__rates__ItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RateItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Item",
        "tags": [
          "rates"
        ]
      }
    },
...
  }
}